requirements.txt
fastapi
pydantic-core>2.0.0
uvicorn